Move some stuff to Resource superclass
